Mock Order History Page:
Concept: For logged-in users, we could add a simple "Order History" page.
Implementation: This page would display a list of mock past orders (e.g., order ID, date, items purchased, total amount). Since we don't have a real backend for orders, this would be simulated data, perhaps generated when a user "completes" a checkout. We could add a link to this page in the user dropdown menu in the header.
Benefit: Adds a common e-commerce feature and fleshes out the "logged-in" experience.